+++ gradle.properties
@@ -1,0 +1,2 @@
+org.gradle.jvmargs=-Dfile.encoding=UTF-8
+org.gradle.console=plain
