+++ src/main/java/racingcar/Application.java
@@ -1,7 +1,7 @@
-package racinggame;
+package racingcar;
 
 public class Application {
     public static void main(String[] args) {
-        // TODO 자동차 경주 게임 구현
+        // TODO: 프로그램 구현
     }
 }
