+++ src/test/java/racingcar/ApplicationTest.java
@@ -1,0 +1,41 @@
+package racingcar;
+
+import camp.nextstep.edu.missionutils.test.NsTest;
+import org.junit.jupiter.api.Test;
+
+import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomNumberInRangeTest;
+import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
+import static org.assertj.core.api.Assertions.assertThat;
+
+class ApplicationTest extends NsTest {
+    private static final int MOVING_FORWARD = 4;
+    private static final int STOP = 3;
+
+    private static final String ERROR_MESSAGE = "[ERROR]";
+
+    @Test
+    void 전진_정지() {
+        assertRandomNumberInRangeTest(
+            () -> {
+                run("pobi,woni", "1");
+                assertThat(output()).contains("pobi : -", "woni : ", "최종 우승자는 pobi 입니다.");
+            },
+            MOVING_FORWARD, STOP
+        );
+    }
+
+    @Test
+    void 이름에_대한_예외_처리() {
+        assertSimpleTest(
+            () -> {
+                runException("pobi,javaji");
+                assertThat(output()).contains(ERROR_MESSAGE);
+            }
+        );
+    }
+
+    @Override
+    public void runMain() {
+        Application.main(new String[]{});
+    }
+}
